feat(datepicker): restructure folders to have wc and react implementations
