+++ packages/web-components/src/components/date-picker/__stories__/date-picker-web-components.mdx
@@ -1,5 +1,5 @@
 import { ArgTypes, Canvas, Markdown, Meta } from '@storybook/addon-docs/blocks';
-import * as DatePickerStories from './date-picker.stories';
+import * as DatePickerStories from './date-picker-web-components.stories';
 
 <Meta of={DatePickerStories} />
