+++ packages/web-components/src/components/date-picker/components/web-components/date-picker-input-skeleton.ts
